+++ data/src/main/java/com/vironit/onlinepharmacy/dao/jdbc/JdbcProductDao.java
@@ -120,12 +120,12 @@ public class JdbcProductDao implements ProductDao {
     }
 
     @Override
-    public int getTotalElements() {
+    public long getTotalElements() {
         String sql = "SELECT COUNT(*) FROM " + PRODUCTS_TABLE;
         try (Connection connection = dataSource.getConnection();
              PreparedStatement preparedStatement = connection.prepareStatement(sql);
              ResultSet resultSet = preparedStatement.executeQuery()) {
-            return resultSet.next() ? resultSet.getInt(1) : -1;
+            return resultSet.next() ? resultSet.getLong(1) : -1;
         } catch (SQLException sqle) {
             throw new DaoException("Error getting total product elements from database", sqle);
         }
